% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_clone.R
\name{add_ml_results}
\alias{add_ml_results}
\title{add_ml_results}
\usage{
add_ml_results(ml_results = NULL, repo = NULL, prefix = NULL)
}
\arguments{
\item{ml_results}{dataframe returned by the ml_dash::run_ml_models() function}

\item{repo}{git_repository object returned by local_clone()}

\item{prefix}{when creating RDS files, the optional prefix will be used; otherwise
it will be set to the github username}
}
\value{
results of the pull request action
}
\description{
Saves a ml_results object to a RDS file and creates a pull request to a
forked mldash project
}
