% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_java.R
\name{check_java}
\alias{check_java}
\title{Checks if Java is installed.}
\usage{
check_java()
}
\value{
TRUE if JAVA_HOME is set.
}
\description{
This function checks to see if the JAVA_HOME environment variable is set. If
not it will check to see if Java is installed on some platforms.
}
\details{
To install Java on Mac (assuming Homebrew is installed):

\if{html}{\out{<div class="sourceCode">}}\preformatted{brew tap adoptopenjdk/openjdk
brew install --cask adoptopenjdk8
}\if{html}{\out{</div>}}

To install Java on Ubuntu run:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sudo apt-get install openjdk-8-jdk
}\if{html}{\out{</div>}}
}
