% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python.R
\name{check_python}
\alias{check_python}
\title{Checks if Python is installed.}
\usage{
check_python()
}
\value{
TRUE if RETICULATE_PYTHON is set.
}
\description{
This function checks to see if the RETICULATE_PYTHON environment variable is set. If
not it will check to see if python is installed on some platforms.
}
