% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Reads in a data frame for a model file}
\usage{
get_data(
  dir = paste0(find.package("mldash"), "/datasets"),
  cache_dir = tempdir(),
  dataname,
  use_cache = TRUE
)
}
\arguments{
\item{dir}{directory containing the dcf files for the datasets.}

\item{cache_dir}{directory where rds data files will be stored.}

\item{use_cache}{whether to read data from the cache if available. If FALSE,
then the data will be retrieved from the \code{data} function parameter.}
}
\value{
the data.frame of the data.
}
\description{
Reads in a data frame for a model file
}
