% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_metrics.R
\name{get_timeseries_metrics}
\alias{get_timeseries_metrics}
\title{Returns time series metrics from the \code{yardstick} package.}
\usage{
get_timeseries_metrics(include_multiclass = FALSE)
}
\value{
a list with time series metric functions from the \code{yardstick} package.
}
\description{
This is a convenience function that will return time series  metrics from the
\code{yardstick} package that can be used with the \code{\link[=run_models]{run_models()}} function.
}
\seealso{
run_models

https://yardstick.tidymodels.org/articles/metric-types.html
}
