% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_clone.R
\name{local_clone}
\alias{local_clone}
\title{local_clone}
\usage{
local_clone(repo_upstream = NULL, verbose = FALSE)
}
\arguments{
\item{repo_upstream}{the  name of the model.}

\item{verbose}{show debug output of the cloning process}
}
\description{
Clone the mldash project to a temp directory
}
