% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldash-package.R
\docType{package}
\name{mldash-package}
\alias{mldash}
\alias{mldash-package}
\title{mldash: Evaluating Machine Learning Models Across Many Datasets}
\description{
Framework for evaluating predictive models across many datasets. Datasets and models are defined in debian control files. Data can be retrieved from many sources and the package will execute and evaluate predictive models using a hold-out validation dataset using common metrics defined in the yardstick package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jbryer.github.io/mldash/}
  \item Report bugs at \url{https://github.com/jbryer/mldash/issues}
}

}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

Other contributors:
\itemize{
  \item Angela Lui \email{angela.lui@cuny.edu} [contributor]
  \item Justin Williams [contributor]
  \item David Simbandumwe [contributor]
  \item Clifton Lee [contributor]
  \item Shane Hylton [contributor]
  \item Joseph Connolly [contributor]
}

}
\keyword{internal}
