% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dataset.R
\name{new_dataset}
\alias{new_dataset}
\title{Create a new data set metadata file.}
\usage{
new_dataset(
  name,
  dir = "./inst/datasets",
  type = c("classification", "regression")[1],
  description = "Description of the dataset",
  source = "The source of the dataset.",
  reference = "APA reference for the dataset.",
  data = "function() {\\n\\t# This function should return a data.frame.\\n\\t}",
  model = "Y ~ X",
  note = "",
  open = interactive(),
  overwrite = FALSE
)
}
\value{
a list with the metadata information.
}
\description{
Create a new data set metadata file.
}
