% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ml_datasets.R
\name{read_ml_datasets}
\alias{read_ml_datasets}
\title{Reads model files for running predictive models.}
\usage{
read_ml_datasets(
  dir = c(paste0(find.package("mldash"), "/datasets")),
  cache_dir = dir,
  pattern = "*.dcf",
  use_cache = TRUE,
  check_for_missing_packages = interactive()
)
}
\arguments{
\item{dir}{directory containing the dcf files for the datasets.}

\item{pattern}{optional regular expression that is used when finding files
to read in. It defaults to all dcf files in the \code{dir}, but could
be a single filename to test a metadata file.}

\item{use_cache}{whether to read data from the cache if available. If FALSE,
then the data will be retrieved from the \code{data} function parameter.}

\item{check_for_missing_packages}{if TRUE you will be prompted to install
missing packages.}

\item{data_cache}{directory where rds data files will be stored.}
}
\value{
a data frame with the following fields:
\itemize{
\item{id}{The filename of the dataset.}
\item{title*}{The name of the dataset from the dcf file.}
\item{type*}{Whether this is for a regression or classification model.}
\item{description}{Description of the dataset.}
\item{source}{The source of the dataset.}
\item{reference}{Reference for the dataset (APA format preferred).}
\item{model*}{The model formula used for the predictive model.}
\item{note}{Any additional information.}
}
\itemize{
\item denotes required fields.
}
}
\description{
Information about models that can be used in the predictive models are stored
in Debian Control Files (dcf). This is the similar to the format used in
RMarkdown YAML (i.e. metadata).
}
\details{
\itemize{
\item{name*}{The name of the dataset.}
\item{type*}{Whether this is for a regression, classification, timeseries, or spatial model.}
\item{description}{Description of the dataset.}
\item{source}{The source of the dataset.}
\item{reference}{Reference for the dataset (APA format preferred).}
\item{data*}{An R function that returns a data.frame.}
\item{model*}{The formula used for the predictive model.}
\item{note}{Any additional information.}
}
\itemize{
\item denotes required fields.
}
}
