% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ml_models.R
\name{read_ml_models}
\alias{read_ml_models}
\title{Reads data files for running predictive models.}
\usage{
read_ml_models(
  dir = paste0(find.package("mldash"), "/models"),
  pattern = "*.dcf",
  check_for_missing_packages = interactive()
)
}
\arguments{
\item{dir}{directory containing the dcf files for the datasets.}

\item{pattern}{optional regular expression that is used when finding files
to read in. It defaults to all dcf files in the \code{dir}, but could
be a single filename to test a metadata file.}

\item{check_for_missing_packages}{if TRUE you will be prompted to install
missing packages.}
}
\value{
a data frame with the following fields:
\itemize{
\item{name}{The name of the dataset.}
\item{type}{Whether this is for a regression or classification model.}
\item{description}{Description of the dataset.}
\item{notes}{Any additional information.}
}
\itemize{
\item denotes required fields.
}
}
\description{
Information about datasets that can be used in the predictive models are stored
in Debian Control Files (dcf). This is the similar to the format used in
RMarkdown YAML (i.e. metadata).
}
\details{
\itemize{
\item{name*}{The name of the dataset.}
\item{type*}{Whether this is for a regression or classification model.}
\item{description}{Description of the dataset.}
\item{train*}{The model formula used for the predictive model.}
\item{predict*}{Any additional information.}
}
\itemize{
\item denotes required fields.
}
}
