% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Runs the predictive models for the given datasets.}
\usage{
run_models(
  datasets,
  models,
  seed,
  training_size = 0.7,
  metrics = get_all_metrics()
)
}
\arguments{
\item{datasets}{the datasets to run the models with. Results from \code{\link[=read_ml_datasets]{read_ml_datasets()}}.}

\item{models}{the models to run. Results from \code{\link[=read_ml_models]{read_ml_models()}}.}

\item{seed}{random seed to set before randomly selecting the training dataset.}

\item{training_size}{the proportion of the data that should be used for training.
The remaining percentage will be used for validation.}

\item{metrics}{list of model performance metrics from the \code{yardstick} package.
See https://yardstick.tidymodels.org/articles/metric-types.html for more information.}
}
\value{
a data.frame with the results of all the models run against all the datasets.
}
\description{
Runs the predictive models for the given datasets.
}
