% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mldash.R
\name{shiny_mldash}
\alias{shiny_mldash}
\title{Run the Loess shiny app.}
\usage{
shiny_mldash(
  model_run_dir = paste0(find.package("mldash"), "/", model_runs),
  ml_results = run_models(ml_datasets, ml_models, seed = 2112),
  ...
)
}
\arguments{
\item{...}{not currently used.}

\item{span_range}{the range of values the user is allowed to set the span.}
}
\description{
This will start the shiny app to demonstrate the loess_vis function.
}
