% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_clone.R
\docType{data}
\name{upstream_owner}
\alias{upstream_owner}
\title{Clones the upstream mldash project, saves a ml_results object and creates a pull request
so others can see past results of run_ml_models().}
\format{
An object of class \code{character} of length 1.
}
\usage{
upstream_owner
}
\description{
Clones the upstream mldash project, saves a ml_results object and creates a pull request
so others can see past results of run_ml_models().
}
\keyword{datasets}
