% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metrics.R
\name{get_all_metrics}
\alias{get_all_metrics}
\title{Returns all metrics from the \code{yardstick} package.}
\usage{
get_all_metrics(include_multiclass = FALSE)
}
\value{
a list with all the metric functions from the \code{yardstick} package.
}
\description{
This is a convenience function that will return all the metrics from the
\code{yardstick} package that can be used with the \code{\link[=run_models]{run_models()}} function.
}
\seealso{
run_models

https://yardstick.tidymodels.org/articles/metric-types.html
}
