% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_model.R
\name{new_model}
\alias{new_model}
\title{Create a new data set metadata file.}
\usage{
new_model(
  filename,
  name,
  type,
  dir = "./inst/models",
  description = "Description of the dataset",
  train_fun = "function(formula, data) {\\n\\t# Run predictive model.\\n\\t}",
  predict_fun =
    "function(model, newdata) {\\n\\t# Return predicted probabilities for classification or nemeric values for regression.\\n\\t}",
  packages = "",
  note = "",
  open = interactive(),
  overwrite = FALSE
)
}
\arguments{
\item{filename}{the name of the file that will be created. This will be the id for the model.}

\item{name}{the  name of the model.}

\item{type}{type of model, either regression (quantitative dependent variables),
or classification (for qualitative dependent variables).}

\item{dir}{the directory to store the model file.}

\item{description}{brief description of the model.}

\item{train_fun}{function that will return a trained model. Must have two parameters:
\code{formula} and \code{data}.}

\item{predict_fun}{function that will return the predicted probabilities for
classification models and predicted Y-values for regression models.}

\item{packages}{comma separated list of any packages that need to be loaded
prior to running either \code{train_fun} or \code{predict_fun}.}

\item{note}{any additional descriptions about the model.}

\item{open}{whether to open the file for editing after it has been created.}

\item{overwrite}{whether to overwrite the model file if it already exists.}
}
\value{
a list with the metadata information.
}
\description{
Create a new data set metadata file.
}
