% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Runs the predictive models for the given datasets.}
\usage{
run_models(
  datasets,
  models,
  seed = sample(1:2^15, 1),
  training_size = 0.7,
  print_errors = FALSE,
  metrics = mldash::get_all_metrics(),
  save_model_fits = FALSE,
  timeout = 60
)
}
\arguments{
\item{datasets}{the datasets to run the models with. Results from \code{\link[=read_ml_datasets]{read_ml_datasets()}}.}

\item{models}{the models to run. Results from \code{\link[=read_ml_models]{read_ml_models()}}.}

\item{seed}{random seed to set before randomly selecting the training dataset.}

\item{training_size}{the proportion of the data that should be used for training.
The remaining percentage will be used for validation.}

\item{print_errors}{if TRUE errors will be printed while the function runs.
Errors will always be saved in the returned object.}

\item{metrics}{list of model performance metrics from the \code{yardstick} package.
See https://yardstick.tidymodels.org/articles/metric-types.html for more information.}

\item{timeout}{the maximum amount of time (in seconds) a model is allowed to run
before it is interrupted, can be \code{Inf} to never expire. Note that not
all models can be interrupted. See \code{\link[R.utils:withTimeout]{R.utils::withTimeout()}} for more information.}
}
\value{
a data.frame with the results of all the models run against all the datasets.
}
\description{
Runs the predictive models for the given datasets.
}
